refactor(luceneQuery): move query builder from composable to util
